requirements.txt
urllib3==1.22
click==6.7
attrdict==2.0.0
ecdsa==0.13
Twisted==17.9.0
requests==2.18.4
IPy==0.83
dnspython==1.15.0
